pyproject.toml
[tool.poetry]
name = "solventgpt"
version = "0.1"
description = "LangChain LLM application for SolventGPT agent"
authors = ["Solvent.Life", "Pedro Lima"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
feedparser = "~6.0.10"
langchain = "~0.1.17"
langchain-openai = "~0.1.4"
langchain-groq = "~0.1.3"
gradio = "^4.28.3"
postlang = "^0.9.0"
yfinance = "^0.2.38"
duckduckgo-search = "^5.3.0b4"
langgraph = "^0.0.40"
fake-useragent = "^1.5.1"
langfuse = "2.6.0"
elevenlabs = "^1.2.1"
python-dateutil = "^2.9.0.post0"
supabase = "^2.4.5"
pyjwt = "^2.8.0"
stripe = "^9.7.0"
markdown = "^3.6"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
